requirements.txt
tensorflow-gpu==2.0.0
keras==2.1.6
numpy~=1.20.3
h5py==2.10.0
pandas~=1.3.5
SQLAlchemy~=1.4.29
numba~=0.54.1